fix a couple of bugs with autogrowth and autoshrinkage discovered while creating new test program for hash tables; other miscellaneous stuff.

+++ _smat_hash.c
@@ -27,7 +27,7 @@ unsigned long
 _smat_hash(hash_table_t *table, db_key_t *key)
 {
   int i, j;
-  unsigned int hash = 0;
+  unsigned long hash = 0;
   void **objects;
   unsigned char *c;
